+++ app/Http/Controllers/EmployeeController.php
@@ -6,9 +6,11 @@ use App\Models\Department;
 use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
+use Spatie\Permission\Models\Role;
 use Yajra\DataTables\DataTables;
 
 class EmployeeController extends Controller
@@ -16,12 +18,19 @@ class EmployeeController extends Controller
     //
     public function index()
     {
+        if (!Auth::user()->can('view_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
         return view('employee.index');
     }
 
     // Employee Data Table
     public function employeeDatatable(Request $request)
     {
+        if (!Auth::user()->can('view_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $employees = User::with('department');
 
         return Datatables::of($employees)
@@ -50,14 +59,29 @@ class EmployeeController extends Controller
                 return null;
             })
             ->addColumn('action', function ($each) {
-                $edit_icon = '<a href="' . route('employee.edit', $each->id) . '" class="text-warning"><i class="fas fa-edit"> </i> </a>';
-                $info_icon = '<a href="' . route('employee.show', $each->id) . '" class="text-primary"><i class="fa fa-info-circle"> </i> </a>';
-                $delete_icon = '<a href="#" class="text-danger delete-btn" data-id="' . $each->id . '"><i class="fa fa-trash-alt"> </i> </a>';
-
+                $edit_icon = '';
+                $info_icon = '';
+                $delete_icon = "";
+                if (Auth::user()->can('edit_employee')) {
+                    $edit_icon = '<a href="' . route('employee.edit', $each->id) . '" class="text-warning"><i class="fas fa-edit"> </i> </a>';
+                }
+                if (Auth::user()->can('view_employee')) {
+                    $info_icon = '<a href="' . route('employee.show', $each->id) . '" class="text-primary"><i class="fa fa-info-circle"> </i> </a>';
+                }
+                if (Auth::user()->can('delete_employee')) {
+                    $delete_icon = '<a href="#" class="text-danger delete-btn" data-id="' . $each->id . '"><i class="fa fa-trash-alt"> </i> </a>';
+                }
                 return '<div class="action-icon">' . $edit_icon . $info_icon . $delete_icon . '</div>';
 
             })
-            ->rawColumns(['profile_img', 'is_present', 'action'])
+            ->addColumn('roles', function ($each) {
+                $output = '';
+                foreach ($each->roles as $role) {
+                    $output .= '<span class="badge rounded-pill badge-primary m-1">' . $role->name . '</span>';
+                }
+                return $output;
+            })
+            ->rawColumns(['profile_img', 'is_present', 'action', 'roles'])
             ->make(true);
 
     }
@@ -65,13 +89,21 @@ class EmployeeController extends Controller
     // Employee Create
     public function create()
     {
+        if (!Auth::user()->can('create_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $departments = Department::orderBy('title', 'asc')->get();
-        return view('employee.create', compact('departments'));
+        $roles = Role::all();
+        return view('employee.create', compact('departments', 'roles'));
     }
 
     // Store Employee data
     public function store(Request $request)
     {
+        if (!Auth::user()->can('create_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
 
         $this->userValidationRule($request, $id = null, 'create');
 
@@ -86,8 +118,9 @@ class EmployeeController extends Controller
             'address' => $request->address,
             'department_id' => $request->department,
             'date_of_join' => $request->date_of_join,
-            'password' => Hash::make($request->password),
             'is_present' => $request->is_present,
+            'pin_code' => $request->pin_code,
+            'password' => Hash::make($request->password),
         ];
         if ($request->hasFile('profileImg')) {
             $profile_image_file = $request->file('profileImg');
@@ -97,7 +130,8 @@ class EmployeeController extends Controller
             $userData['image'] = $profile_image_name;
         }
 
-        User::create($userData);
+        $user = User::create($userData);
+        $user->syncRoles($request->roles);
 
         return redirect()->route('employee.index')->with(['createSuccess' => 'Employee Successfully create']);
     }
@@ -105,13 +139,22 @@ class EmployeeController extends Controller
     // Employee Edit
     public function edit($id)
     {
+        if (!Auth::user()->can('edit_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $departments = Department::orderBy('title', 'asc')->get();
         $employee = User::findOrFail($id);
-        return view('employee.edit', compact('departments', 'employee'));
+        $old_roles = $employee->roles->pluck('id')->toArray();
+        $roles = Role::all();
+        return view('employee.edit', compact('departments', 'employee', 'old_roles', 'roles'));
     }
     // Employee Update
     public function update($id, Request $request)
     {
+        if (!Auth::user()->can('edit_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
 
         $this->userValidationRule($request, $id, 'update');
 
@@ -127,6 +170,7 @@ class EmployeeController extends Controller
             'department_id' => $request->department,
             'date_of_join' => $request->date_of_join,
             'is_present' => $request->is_present,
+            'pin_code' => $request->pin_code,
         ];
         $employee = User::where("id", $id)->first();
         if ($request->password != null) {
@@ -134,6 +178,7 @@ class EmployeeController extends Controller
         } else {
             $data['password'] = $employee->password;
         }
+
         if ($request->hasFile('profileImg')) {
 
             Storage::disk('public')->delete('employee/' . $employee->image);
@@ -145,12 +190,19 @@ class EmployeeController extends Controller
         }
 
         User::where('id', $request->id)->update($data);
+        $user = User::findOrFail($request->id);
+        $user->syncRoles($request->roles);
+
         return redirect()->route('employee.index')->with(['createSuccess' => 'Employee Successfully Update']);
     }
 
     // User Details
     public function show($id)
     {
+        if (!Auth::user()->can('view_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $employee = User::findOrFail($id);
         return view('employee.show', compact('employee'));
     }
@@ -158,6 +210,10 @@ class EmployeeController extends Controller
     // Delete Employee
     public function destroy($id)
     {
+        if (!Auth::user()->can('delete_employee')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         User::findOrFail($id)->delete();
         return 'success';
     }
@@ -182,24 +238,5 @@ class EmployeeController extends Controller
         $validationRule['password'] = $action == 'create' ? 'required' : '';
         Validator::make($request->all(), $validationRule)->validate();
 
-    }
-
-    // User Data
-    private function userData($request)
-    {
-        return [
-            'name' => $request->name,
-            'email' => $request->email,
-            'employee_id' => $request->employeeId,
-            'phone' => $request->phone,
-            'nrc_number' => $request->nrcNumber,
-            'birthday' => $request->birthday,
-            'gender' => $request->gender,
-            'address' => $request->address,
-            'department_id' => $request->department,
-            'date_of_join' => $request->date_of_join,
-            'is_present' => $request->is_present,
-
-        ];
     }
 }
