+++ app/Http/Controllers/PermissionController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 use Spatie\Permission\Models\Permission;
 use Yajra\DataTables\DataTables;
@@ -13,25 +13,41 @@ class PermissionController extends Controller
     //
     public function index()
     {
+        if (!Auth::user()->can('view_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
         return view('permission.index');
     }
 
     // Permission DataTable
     public function permissionDatatable(Request $request)
     {
+        if (!Auth::user()->can('view_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
         $permissions = Permission::query();
 
         return DataTables::of($permissions)
+            ->editColumn('created_at', function ($each) {
+                return $each->created_at->format('Y-m-d H:i:s');
+            })
             ->editColumn('updated_at', function ($each) {
-                return Carbon::parse($each->updated_at)->format('Y-md H:i:s');
+                return $each->updated_at->format('Y-m-d H:i:s');
             })
             ->addColumn('plus-icon', function ($each) {
                 return null;
             })
             ->addColumn('action', function ($each) {
-                $edit_icon = '<a href="' . route('permission.edit', $each->id) . '" class="text-warning"><i class="fas fa-edit"> </i> </a>';
+                $edit_icon = '';
+                $delete_icon = '';
 
-                $delete_icon = '<a href="#" class="text-danger delete-btn" data-id="' . $each->id . '"><i class="fa fa-trash-alt"> </i> </a>';
+                if (Auth::user()->can('edit_permission')) {
+                    $edit_icon = '<a href="' . route('permission.edit', $each->id) . '" class="text-warning"><i class="fas fa-edit"> </i> </a>';
+                }
+
+                if (Auth::user()->can('delete_permission')) {
+                    $delete_icon = '<a href="#" class="text-danger delete-btn" data-id="' . $each->id . '"><i class="fa fa-trash-alt"> </i> </a>';
+                }
 
                 return '<div class="action-icon">' . $edit_icon . $delete_icon . '</div>';
 
@@ -43,12 +59,20 @@ class PermissionController extends Controller
     // Permission Create
     public function create()
     {
+        if (!Auth::user()->can('create_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('permission.create');
     }
 
     // Store Permission data
     public function store(Request $request)
     {
+        if (!Auth::user()->can('create_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $this->validationRule($request);
         $permission = [
             'name' => $request->name,
@@ -62,12 +86,20 @@ class PermissionController extends Controller
     // Permission Edit
     public function edit($id)
     {
+        if (!Auth::user()->can('edit_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $permission = Permission::findOrFail($id);
         return view('permission.edit', compact('permission'));
     }
     // Permission Update
     public function update($id, Request $request)
     {
+        if (!Auth::user()->can('edit_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $this->validationRule($request);
         $data = [
             'name' => $request->name,
@@ -80,6 +112,10 @@ class PermissionController extends Controller
     // Delete Permission
     public function destroy($id)
     {
+        if (!Auth::user()->can('delete_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         Permission::findOrFail($id)->delete();
         return 'success';
     }
